+++ NEWS.md
@@ -5,6 +5,8 @@
 - improve evaluation of linear predictors in `theme` function (issue #8)
 - remove dependency on `expm` package (issue #10)
 - preliminary integration of code from SCnext/mixedSCGLR written by Jocelyn Chauvet (issue #11)
+- error message for lack of convergence were improved and crossval no longer
+  stops when a fold didn't converge (issue #12)
 
 # version 3.0
 This major version introduces a new feature allowing to group covariates in so called **themes**.